requirements.txt
# Insurance Graph RAG System Requirements

# Core dependencies
python-dotenv>=1.0.0
openai>=1.0.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.5.0

# NLP and embeddings
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.2
tiktoken>=0.5.0
spacy>=3.5.0

# Document processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=0.8.11
beautifulsoup4>=4.11.0
requests>=2.28.0

# Graph database
networkx>=3.0

# Web and parsing
regex>=2022.10.31

# Testing
unittest-xml-reporting>=3.2.0

# Optional: Download spaCy model
# python -m spacy download en_core_web_lg